MotionCapture now works
run MotionCapture to get a list of images (50x50) that make up the
difference ready to be transmitted to the display. I think there may be
problems with the coordinates used in gathering the displays, though.
Because some of the images are straight black... leading me to believe
the images are constructed backwords (with larger coordinates first)