Finally found and fixed problem
This is a scaled back version. I deleted a lot of advanced functions
until I could more easily troubleshoot my problems. I also created a
class BoundingBox to help standardize creating coordinates everywhere.
But that wasn't the error. I found that finding the box that represented
changes in the difference image was wrong. 1) it used "if True in
topSlice:" which does not return True if there are positive values in
the slice. It appears that it only returns True when a majority of
values are positive. 2) bottom and right coordinates are off by one
pixel. This means small, 1x5 images (or similar) are actually
interpreted as 0x4, creating an empty image that then causes errors past
that. Both errors have been corrected. 1) I used if sum(topSlice): 2) if
coordinate contains a positive pixel, I add +1 to coordinate and break
from for-loop. I hope to re-implement the advanced functions again.